refactor(archive): move replayer logic into library module
