+++ tools/archive/src/bin/replayer.rs
@@ -1,23 +1,7 @@
-use archive::read::ArchiveReader;
-use shared::clap;
+use archive::replayer::{self, Args};
 use shared::clap::Parser;
-use shared::protobuf::event::{event::PeerObserverEvent, Event};
-use shared::{log, simple_logger};
-use std::io::ErrorKind;
-use std::path::PathBuf;
+use shared::simple_logger;
 use std::process::ExitCode;
-
-#[derive(Parser, Debug)]
-#[command(version, about = "Read and display peer-observer archive files")]
-struct Args {
-    /// Archive files to read.
-    #[arg(value_name = "FILE", required = true)]
-    files: Vec<PathBuf>,
-
-    /// The log level the tool should run on.
-    #[arg(short, long, default_value_t = log::Level::Info)]
-    log_level: log::Level,
-}
 
 fn main() -> ExitCode {
     let args = Args::parse();
@@ -26,91 +10,9 @@ fn main() -> ExitCode {
         eprintln!("replayer tool error: {}", e);
     }
 
-    let mut had_error = false;
-
-    for path in &args.files {
-        if args.files.len() > 1 {
-            log::info!("=== {} ===", path.display());
-        }
-
-        let archive = match ArchiveReader::open(path) {
-            Ok(archive) => archive,
-            Err(e) => {
-                log::error!("failed to read archive {}: {e}", path.display());
-                had_error = true;
-                continue;
-            }
-        };
-
-        log::info!("header: {}", archive.header);
-
-        let mut events_count = 0;
-        for (idx, event) in archive.enumerate() {
-            let n = idx + 1;
-            let event = match event {
-                Ok(event) => event,
-                Err(e) if e.kind() == ErrorKind::UnexpectedEof => {
-                    log::error!(
-                        "archive {} ended unexpectedly while reading event {n}: {e}",
-                        path.display()
-                    );
-                    had_error = true;
-                    break;
-                }
-                Err(e) => {
-                    log::error!(
-                        "failed to read archive {} at event {n}: {e}",
-                        path.display()
-                    );
-                    had_error = true;
-                    break;
-                }
-            };
-
-            if let Err(e) = display_event(n, &event) {
-                log::warn!(
-                    "failed to display archive {} event {n}: {e}",
-                    path.display()
-                );
-            }
-            events_count = n;
-        }
-
-        log::info!("total: {} events", events_count);
-    }
-
-    if had_error {
+    if replayer::run(&args) {
         ExitCode::FAILURE
     } else {
         ExitCode::SUCCESS
     }
-}
-
-fn display_event(n: usize, event: &Event) -> Result<(), &'static str> {
-    let ts = event.timestamp;
-    match &event.peer_observer_event {
-        Some(PeerObserverEvent::EbpfExtractor(e)) => {
-            let ebpf_event = e.ebpf_event.as_ref().ok_or("missing ebpf event")?;
-            log::info!("[{n}] ts={ts} ebpf: {ebpf_event}");
-        }
-        Some(PeerObserverEvent::RpcExtractor(r)) => {
-            let rpc_event = r.rpc_event.as_ref().ok_or("missing rpc event")?;
-            log::info!("[{n}] ts={ts} rpc: {rpc_event}");
-        }
-        Some(PeerObserverEvent::P2pExtractor(p)) => {
-            let p2p_event = p.p2p_event.as_ref().ok_or("missing p2p event")?;
-            log::info!("[{n}] ts={ts} p2p: {p2p_event}");
-        }
-        Some(PeerObserverEvent::LogExtractor(l)) => {
-            let log_event = l.log_event.as_ref().ok_or("missing log event")?;
-            log::info!("[{n}] ts={ts} log: {log_event}");
-        }
-        Some(PeerObserverEvent::IpcExtractor(i)) => {
-            let ipc_event = i.ipc_event.as_ref().ok_or("missing ipc event")?;
-            log::info!("[{n}] ts={ts} ipc: {ipc_event}");
-        }
-        None => log::info!("[{n}] ts={ts} <unknown>"),
-    }
-
-    Ok(())
 }
