+++ tools/archive/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod archiver;
 
 pub mod read;
+
+pub mod replayer;
